Backend/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
Pillow==10.3.0
django-cors-headers==4.3.1